cloud-vision/requirements.txt
protobuf
google-cloud-storage
google-cloud-vision